chore(module:code-editor): refactor (#4719)
