+++ components/code-editor/code-editor.spec.ts
