+++ components/code-editor/typings.ts
